Upload handler changed to int64.

+++ common/src/java/be/telio/mediastore/ui/upload/OutputStreamListener.java
@@ -18,7 +18,7 @@ package be.telio.mediastore.ui.upload;
 public interface OutputStreamListener
 {
     public void start();
-    public void bytesRead(int bytesRead);
+    public void bytesRead(long bytesRead);
     public void error(String message);
     public void done();
 }
